switch reading meta.json to `cosa meta` command
Since the builds are happening on remote builders we need to filter
access through `cosa meta` which knows how to proxy the command over to
the remote builder to read the file there rather than acutally in
the x86_64 pod we are running in on OpenShift.

We also update the build.Jenkinsfile job here for consistency and
rework that block so it doesn't have to also read builds.json.