bump-lockfile: drop use of cosa fetch
When we dropped the fedora-bootc submodule [1] it actually caused
some issues with bump-lockfile exposing the fact that we were actually
still using the submodule (even though it was out of date, because
we thought it wasn't being used) for inputs for `cosa fetch` to update
the lockfile.

Since we are no longer using the submodule (and no longer have the
submodule) we need to bump lockfiles a different way. The best thing
I can think to do is to just do a full container build first and then
do the logic with the lockfiles generated from that exchange.

[1] https://github.com/coreos/fedora-coreos-config/pull/3921